backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
langchain>=0.1.10
langchain-openai>=0.0.8
langchain-community>=0.0.25
python-dotenv>=1.0.1
requests>=2.31.0
